refactor: outsource login status bar component to be reusable
